+++ packages/analytics/analytics-chart/sandbox/App.vue
@@ -537,7 +537,6 @@ body {
   min-height: 100%;
 }
 .sandbox-container {
-  min-height: 540px;
   margin: $spacing-md;
   padding: $spacing-md;